requirements.txt
# requirements.txt

en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
python-telegram-bot==21.4
scikit-learn==1.5.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
